Immediately update UI on item updates instead of waiting for backend
Instead of re-fetching the entire list after each item update the effect of adding/updating/deleting an item is now immediately reflected in the UI. This is done by updating the app state when a request is sent instead of waiting for the request to complete. If a request fails, then the UI changes will also reverted by updating the app state appropriately.

Remains to indicate somehow that an item update has not yet been persisted.